refactored out middleware to its own folder
