+++ router/api/index.js
@@ -1,6 +1,6 @@
-var express = require('express')
-var passport = require('../auth/passport')
-var User = require('../../models/User.js')
+const express = require('express')
+const passport = require('../auth/passport')
+const addLastActivity = require('../../middleware/add-last-activity')
 
 module.exports = function(app, sessionStore) {
   console.log('API module')
@@ -8,26 +8,6 @@ module.exports = function(app, sessionStore) {
   const io = require('./socket-server')(app)
 
   const router = new express.Router()
-
-  app.use(function(req, res, next) {
-    if (req.user) {
-      const { id, lastActivityAt } = req.user
-      const todaysDateInMS = Date.now()
-      const oneDayElapsed = 1000 * 60 * 60 * 24
-      const lastActivityInMS = new Date(lastActivityAt).getTime()
-
-      if (lastActivityInMS + oneDayElapsed <= todaysDateInMS) {
-        const todaysDateFormatted = new Date(todaysDateInMS)
-        User.updateOne({ _id: id }, { lastActivityAt: todaysDateFormatted })
-          .then(() => next())
-          .catch(err => next(err))
-      } else {
-        next()
-      }
-    } else {
-      next()
-    }
-  })
 
   require('./volunteers')(router)
   require('./user')(router)
@@ -39,5 +19,6 @@ module.exports = function(app, sessionStore) {
   require('./sockets')(io, sessionStore)
   require('./moderate')(router)
 
+  app.use(addLastActivity)
   app.use('/api', passport.isAuthenticated, router)
 }
